# user-service/docs/openapi.yaml
components:
    schemas:
...
        Code:
            enum:
                - 0
                - 10000
                - 10001
                - 20000
                - 20001
                - 20002
                - 20003
                - 20004
                - 20005
                - 30000
                - 40000
                - 50000
                - 90000
            type: integer
            x-enum-varnames:
                - CodeOK
                - CodeBadRequest
                - CodeValidationFailed
                - CodeUnauthenticated
                - CodeTokenInvalid
                - CodeTokenExpired
                - CodeTokenRevoked
                - CodeMFARequired
                - CodeUserAccountLocked
                - CodeForbidden
                - CodeConflict
                - CodeNotFound
                - CodeInternalError
...
        CreateUserRequest:
            properties:
                nickname:
                    example: Alice
                    maxLength: 128
                    minLength: 1
                    type: string
                password:
                    example: secret
                    minLength: 1
                    type: string
                status:
                    allOf:
                        - $ref: '#/components/schemas/UserStatus'
                    example: 100
                username:
                    example: alice
                    maxLength: 255
                    minLength: 1
                    type: string
            required:
                - nickname
                - password
                - username
            type: object
...
        Envelope:
            properties:
                code:
                    $ref: '#/components/schemas/Code'
                data: {}
                errors: {}
                message:
                    type: string
                success:
                    type: boolean
            type: object
...
        Pagination:
            properties:
                has_next:
                    example: true
                    type: boolean
                next_cursor:
                    example: 0190c8d2-8d8a-7a01-9f43-0f91fb4e2b7c
                    type: string
                page_size:
                    example: 50
                    type: integer
            type: object
...
        UserListResponseDoc:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/UserResponseDoc'
                    type: array
                pagination:
                    $ref: '#/components/schemas/Pagination'
            type: object
        UserResponseDoc:
            properties:
                created_at:
                    example: 1.7802888e+12
                    type: integer
                nickname:
                    example: Alice
                    type: string
                status:
                    example: 100
                    type: integer
                updated_at:
                    example: 1.7802888e+12
                    type: integer
                user_id:
                    example: 018f6f3e-7c4d-7b2a-9f8a-4f6b1b2c3d4e
                    type: string
                username:
                    example: alice
                    type: string
            type: object
        UserStatus:
            enum:
                - 100
                - 200
                - 300
            format: int64
            type: integer
            x-enum-varnames:
                - UserStatusNormal
                - UserStatusDisabled
                - UserStatusMustChangePassword
    securitySchemes:
        BearerAuth:
            bearerFormat: JWT
            description: 输入 Bearer token，格式为：Bearer <token>
            scheme: bearer
            type: http
info:
    contact: {}
    description: AegisCore 用户服务 API 文档，覆盖认证会话、用户资料、角色管理、权限目录、RBAC 授权保护的业务接口和服务健康检查。
    title: AegisCore User Services API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /users:
        get:
            description: 基于 user_id keyset 分页查询用户资料列表，支持按用户昵称、用户名和用户状态过滤，默认排除软删除用户。page_size 未传或小于 1 时使用默认值 10，超过 100 时按 100 处理。
            parameters:
                - description: 分页游标，上一页最后一条用户的 user_id
                  in: query
                  name: cursor
                  schema:
                    type: string
                - description: 每页数量，未传或小于 1 时默认为 10，超过 100 时按 100 处理
                  in: query
                  name: page_size
                  schema:
                    maximum: 100
                    minimum: 1
                    type: integer
                - description: 用户昵称模糊匹配
                  in: query
                  name: nickname
                  schema:
                    type: string
                - description: 用户名精确匹配
                  in: query
                  name: username
                  schema:
                    type: string
                - description: 用户状态：100 正常，200 冻结/停用，300 必须修改密码
                  in: query
                  name: status
                  schema:
                    enum:
                        - 100
                        - 200
                        - 300
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/UserListResponseDoc'
                                      type: object
                    description: 查询成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 查询参数错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 分页查询用户列表
            tags:
                - 用户
        post:
            description: 创建一个新的用户资料。请求体使用共享校验器校验，用户名必须唯一，status 缺省为 100。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequest'
                description: 创建用户请求
                required: true
                x-originalParamName: request
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/UserResponseDoc'
                                      type: object
                    description: 创建成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求体错误或参数校验失败
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 用户已存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 创建用户
            tags:
                - 用户
    /users/{user_id}:
        get:
            description: 通过外部 UUID 用户 ID 查询用户基础资料。
            parameters:
                - description: 用户ID
                  in: path
                  name: user_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/UserResponseDoc'
                                      type: object
                    description: 查询成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 用户 ID 参数错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 用户不存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 查询用户资料
            tags:
                - 用户
...
servers:
    - url: /api/v1